Extract SimpleBakeTask
HelperTask and BehaviorTask were so similar that their differences could
be expressed as simple set of properties.
